Move API search into separate class, add new cmd
