+++ ckanext/attribution/commands/migration/__init__.py
@@ -1,3 +1,4 @@
 from .parser import Parser
 from .common import multi_choice, rgx
 from .combiner import Combiner
+from .api_search import APISearch
